move: ViewModelを一箇所にまとめ
